# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for EduCRM: JWT auth (access + refresh), users, teachers, rooms, groups, schedules, attendance, grades, payments, files, notifications, dashboard, and AI analytics. Success responses use a JSON envelope: `{\"success\":true,\"data\":...}`. Errors use `{\"success\":false,\"error\":{\"code\",\"message\",\"kind\"}}`.",
        "title": "EduCRM API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@educrm.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticates a user by email (if login contains @) or phone and returns JWT access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with email or phone",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.TokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "description": "Revokes refresh tokens using either Authorization Bearer access token (all sessions) or refresh_token in body (single session).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer access_token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Optional refresh_token",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.MessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user profile (requires access token).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.CurrentUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Exchanges a valid refresh token for a new access/refresh pair (rotation).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.TokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Headline counts (UTC date for \"today\"). Debtors = distinct students with unpaid/overdue rows for the current billing month (month_for). Today lessons = schedule slots on today's weekday for active groups whose date range includes today. Today payments = rows with payment_date = today (UTC) or, if null, created_at date = today. Revenue = sum of amount_minor for paid_full/paid_partial, non-free payments created in the selected UTC month (default current).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dashboard statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference instant RFC3339 (default: now UTC)",
                        "name": "as_of",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Revenue month YYYY-MM (UTC, default: current month)",
                        "name": "year_month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.DashboardSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.CurrentUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.DashboardCountsResponse": {
            "type": "object",
            "properties": {
                "active_groups": {
                    "type": "integer"
                },
                "debtors": {
                    "type": "integer"
                },
                "today_lessons": {
                    "type": "integer"
                },
                "today_payments": {
                    "type": "integer"
                },
                "total_students": {
                    "type": "integer"
                },
                "total_teachers": {
                    "type": "integer"
                }
            }
        },
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.DashboardMonthlyRevenueResponse": {
            "type": "object",
            "properties": {
                "payment_count": {
                    "type": "integer"
                },
                "total_amount_minor": {
                    "type": "integer"
                },
                "year_month": {
                    "type": "string"
                }
            }
        },
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.DashboardSummaryResponse": {
            "type": "object",
            "properties": {
                "counts": {
                    "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.DashboardCountsResponse"
                },
                "revenue": {
                    "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.DashboardMonthlyRevenueResponse"
                }
            }
        },
...
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.LoginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "teacher@school.edu"
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8,
                    "example": "your-secure-password"
                }
            }
        },
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.LogoutRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "opaque-refresh-token"
                }
            }
        },
...
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_educrm_educrm-backend_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.ErrorBody"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_educrm_educrm-backend_pkg_response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}